# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PubWIFIprot API",
        "description": "API documentation for the PubWIFIprot backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pubwifiprot.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard statistics",
                "operationId": "a4e953abeda97cdeaa1c2af786aea1f3",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "active_today": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "new_this_week": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "networks": {
                                            "properties": {
                                                "total_reported": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "pending_review": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "confirmed_malicious": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "domains": {
                                            "properties": {
                                                "total_malicious": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "blocked_today": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "new_this_week": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "sessions": {
                                            "properties": {
                                                "active": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "total_today": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "average_duration": {
                                                    "type": "integer",
                                                    "example": 45
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "recent_activity": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "new_user",
                                                            "network_report",
                                                            "malicious_domain",
                                                            "session_start"
                                                        ]
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}